+++ packages/react-components/react-components/src/index.ts
@@ -34,6 +34,7 @@ export {
   useArrowNavigationGroup,
   useFocusableGroup,
   useFocusFinders,
+  useFocusWithin,
   useKeyboardNavAttribute,
   useModalAttributes,
 } from '@fluentui/react-tabster';
@@ -90,6 +91,7 @@ export {
   getNativeElementProps,
   getPartitionedNativeProps,
   getSlots,
+  IdPrefixProvider,
   resetIdsForTests,
   resolveShorthand,
   SSRProvider,
@@ -196,7 +198,10 @@ export {
 export type {
   AvatarNamedColor,
   AvatarProps,
+  // AvatarSizes is deprecated but removing it would be a breaking change
+  // eslint-disable-next-line deprecation/deprecation
   AvatarSizes,
+  AvatarSize,
   AvatarSlots,
   AvatarState,
   AvatarGroupProps,