+++ packages/react-components/react-datepicker/stories/CalendarGrid/CalendarGridDefault.stories.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { CalendarGrid, CalendarGridProps } from '@fluentui/react-datepicker';
-
-export const Default = (props: Partial<CalendarGridProps>) => <CalendarGrid {...props} />;