+++ packages/react-components/react-tree/src/components/TreeItem/renderTreeItem.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { getSlots } from '@fluentui/react-utilities';
 import type { TreeItemState, TreeItemSlots, TreeItemContextValues } from './TreeItem.types';
-import { TreeItemProvider } from '../../contexts/treeItemContext';
+import { TreeItemProvider } from '../../contexts';
+import { ButtonContextProvider } from '@fluentui/react-button';
 
 /**
  * Render the final JSX of TreeItem
@@ -11,13 +12,16 @@ export const renderTreeItem_unstable = (state: TreeItemState, contextValues: Tre
 
   return (
     <TreeItemProvider value={contextValues.treeItem}>
-      <slots.groupper {...slotProps.groupper}>
-        <slots.root {...slotProps.root}>
+      <slots.root {...slotProps.root}>
+        <slots.content {...slotProps.content}>
           {slots.expandIcon && <slots.expandIcon {...slotProps.expandIcon} />}
-          {slotProps.root.children}
-          {slots.actions && <slots.actions {...slotProps.actions} />}
-        </slots.root>
-      </slots.groupper>
+          {slotProps.content.children}
+          <ButtonContextProvider value={contextValues.button}>
+            {slots.actions && <slots.actions {...slotProps.actions} />}
+          </ButtonContextProvider>
+        </slots.content>
+        {slots.subtree && <slots.subtree {...slotProps.subtree} />}
+      </slots.root>
     </TreeItemProvider>
   );
 };