+++ packages/react-cards/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @fluentui/react-cards
 
-This log was last generated on Wed, 18 Jan 2023 07:42:14 GMT and should not be manually modified.
+This log was last generated on Mon, 30 Jan 2023 07:36:22 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.205.47](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v0.205.47)
+
+Mon, 30 Jan 2023 07:36:22 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-cards_v0.205.46..@fluentui/react-cards_v0.205.47)
+
+### Patches
+
+- Bump @fluentui/react to v8.105.0 ([PR #26522](https://github.com/microsoft/fluentui/pull/26522) by beachball)
+
+## [0.205.46](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v0.205.46)
+
+Fri, 27 Jan 2023 07:37:51 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-cards_v0.205.45..@fluentui/react-cards_v0.205.46)
+
+### Patches
+
+- Bump @fluentui/react to v8.104.8 ([PR #26520](https://github.com/microsoft/fluentui/pull/26520) by beachball)
+- Bump @fluentui/foundation-legacy to v8.2.27 ([PR #26520](https://github.com/microsoft/fluentui/pull/26520) by beachball)
+
+## [0.205.45](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v0.205.45)
+
+Thu, 26 Jan 2023 07:35:32 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-cards_v0.205.44..@fluentui/react-cards_v0.205.45)
+
+### Patches
+
+- Bump @fluentui/react to v8.104.7 ([PR #26509](https://github.com/microsoft/fluentui/pull/26509) by beachball)
 
 ## [0.205.44](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v0.205.44)